+++ bin/remote-prover/src/server/status.rs
@@ -1,8 +1,5 @@
-use proto::worker_status_api_server::WorkerStatusApiServer;
-use tonic::{Request, Response, Status};
+use miden_node_proto::generated as grpc;
 
-use crate::generated::worker_status_api_server::WorkerStatusApi;
-use crate::generated::{self as proto};
 use crate::server::proof_kind::ProofKind;
 
 pub struct StatusService {
@@ -10,17 +7,28 @@ pub struct StatusService {
 }
 
 impl StatusService {
-    pub fn new(kind: ProofKind) -> WorkerStatusApiServer<Self> {
-        WorkerStatusApiServer::new(Self { kind })
+    pub fn new(kind: ProofKind) -> Self {
+        Self { kind }
     }
 }
 
-#[async_trait::async_trait]
-impl WorkerStatusApi for StatusService {
-    async fn status(&self, _: Request<()>) -> Result<Response<proto::WorkerStatus>, Status> {
-        Ok(Response::new(proto::WorkerStatus {
+#[tonic::async_trait]
+impl grpc::server::remote_prover_worker_status_api::Status for StatusService {
+    type Input = ();
+    type Output = grpc::remote_prover::WorkerStatus;
+
+    async fn handle(&self, _input: Self::Input) -> tonic::Result<Self::Output> {
+        Ok(grpc::remote_prover::WorkerStatus {
             version: env!("CARGO_PKG_VERSION").to_string(),
             supported_proof_type: self.kind as i32,
-        }))
+        })
+    }
+
+    fn decode(_request: ()) -> tonic::Result<Self::Input> {
+        Ok(())
+    }
+
+    fn encode(output: Self::Output) -> tonic::Result<grpc::remote_prover::WorkerStatus> {
+        Ok(output)
     }
 }